revert ptr move
