+++ tests/interpreters/test_ml_program_interpreter.py
@@ -11,8 +11,8 @@ from xdsl.dialects.builtin import (
 )
 from xdsl.interpreter import Interpreter
 from xdsl.interpreters.ml_program import MLProgramFunctions
+from xdsl.interpreters.ptr import TypedPtr
 from xdsl.interpreters.shaped_array import ShapedArray
-from xdsl.interpreters.utils.ptr import TypedPtr
 
 
 def test_ml_program_global_load_constant():
