+++ xdsl/interpreters/memref.py
@@ -10,8 +10,8 @@ from xdsl.interpreter import (
     register_impls,
 )
 from xdsl.interpreters.builtin import xtype_for_el_type
+from xdsl.interpreters.ptr import TypedPtr
 from xdsl.interpreters.shaped_array import ShapedArray
-from xdsl.interpreters.utils.ptr import TypedPtr
 from xdsl.ir import Attribute
 from xdsl.traits import SymbolTable
